+++ app/src/main/java/vn/nhantd/mycareer/model/user/Salary.java
@@ -4,8 +4,8 @@ import java.io.Serializable;
 
 public class Salary implements Serializable {
     private String currency_unit = "";
-    private Double from_salary = 0d;
-    private Double to_salary = 0d;
+    private Long from_salary = 0L;
+    private Long to_salary = 0L;
 
     // Standard getters & setters
     public String getCurrency_unit() {
@@ -16,19 +16,19 @@ public class Salary implements Serializable {
         this.currency_unit = currency_unit;
     }
 
-    public Double getFrom_salary() {
+    public Long getFrom_salary() {
         return from_salary;
     }
 
-    public void setFrom_salary(Double from_salary) {
+    public void setFrom_salary(Long from_salary) {
         this.from_salary = from_salary;
     }
 
-    public Double getTo_salary() {
+    public Long getTo_salary() {
         return to_salary;
     }
 
-    public void setTo_salary(Double to_salary) {
+    public void setTo_salary(Long to_salary) {
         this.to_salary = to_salary;
     }
